Move layout utility back into encoder crate
