+++ crates/abi/encoder/src/types/numeric.rs
@@ -13,7 +13,7 @@ where
         let type_is_signed = T::min_value() < T::zero();
 
         let payload_size = self.byte_size() - 1;
-        let layout_marker = layout::integer(
+        let layout_marker = layout_integer(
             T::max_byte_size() as u32 - 1,
             payload_size as u32,
             type_is_signed,
@@ -87,4 +87,36 @@ impl Numeric for i64 {
 
 impl Numeric for u64 {
     type Unsigned = u64;
+}
+
+pub fn layout_integer(max_width_in_bytes: u32, width_in_bytes: u32, signed: bool) -> u8 {
+    use svm_abi_layout::layout::*;
+    debug_assert!(width_in_bytes <= max_width_in_bytes);
+    debug_assert!(max_width_in_bytes <= 8);
+    match (max_width_in_bytes, (width_in_bytes - 1) as u8, signed) {
+        (1, 0, true) => I8,
+        (1, 0, false) => U8,
+        (2, n, true) => I16_1B | ((n & 0b1) << 4),
+        (2, n, false) => U16_1B | ((n & 0b1) << 4),
+        (4, n, true) => I32_1B | ((n & 0b11) << 4),
+        (4, n, false) => U32_1B | ((n & 0b11) << 4),
+        (8, n, true) => I64_1B | ((n & 0b111) << 4),
+        (8, n, false) => U64_1B | ((n & 0b111) << 4),
+        _ => panic!("Invalid argument for layout information."),
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use svm_abi_layout::layout;
+
+    #[test]
+    fn integer_layouts() {
+        assert_eq!(layout_integer(1, 1, true), layout::I8);
+        assert_eq!(layout_integer(1, 1, false), layout::U8);
+        assert_eq!(layout_integer(2, 1, true), layout::I16_1B);
+        assert_eq!(layout_integer(2, 2, false), layout::U16_2B);
+        assert_eq!(layout_integer(8, 8, false), layout::U64_8B);
+    }
 }
